LPS-91804 Move rules to versioned package
