+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/resource/v1_0/DataRecordResourceImpl.java
@@ -22,9 +22,9 @@ import com.liferay.data.engine.rest.dto.v1_0.DataDefinitionRule;
 import com.liferay.data.engine.rest.dto.v1_0.DataRecord;
 import com.liferay.data.engine.rest.internal.dto.v1_0.util.DataDefinitionUtil;
 import com.liferay.data.engine.rest.internal.dto.v1_0.util.DataRecordValueUtil;
-import com.liferay.data.engine.rest.internal.rule.DataRuleFunction;
-import com.liferay.data.engine.rest.internal.rule.DataRuleFunctionFactory;
-import com.liferay.data.engine.rest.internal.rule.DataRuleFunctionResult;
+import com.liferay.data.engine.rest.internal.rule.v1_0.DataRuleFunction;
+import com.liferay.data.engine.rest.internal.rule.v1_0.DataRuleFunctionFactory;
+import com.liferay.data.engine.rest.internal.rule.v1_0.DataRuleFunctionResult;
 import com.liferay.data.engine.rest.internal.storage.DataStorage;
 import com.liferay.data.engine.rest.resource.v1_0.DataRecordResource;
 import com.liferay.dynamic.data.lists.model.DDLRecord;
