+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/rule/v1_0/IntegerLiteralDataRuleFunction.java
@@ -12,17 +12,18 @@
  * details.
  */
 
-package com.liferay.data.engine.rest.internal.rule;
+package com.liferay.data.engine.rest.internal.rule.v1_0;
 
 import com.liferay.data.engine.constants.DataDefinitionRuleConstants;
 import com.liferay.data.engine.rest.dto.v1_0.DataDefinitionField;
 import com.liferay.data.engine.rest.dto.v1_0.DataDefinitionRuleParameter;
-import com.liferay.portal.kernel.util.Validator;
+
+import org.apache.commons.lang.math.NumberUtils;
 
 /**
  * @author Jeyvison Nascimento
  */
-public class URLDataRuleFunction implements DataRuleFunction {
+public class IntegerLiteralDataRuleFunction implements DataRuleFunction {
 
 	@Override
 	public DataRuleFunctionResult validate(
@@ -35,14 +36,21 @@ public class URLDataRuleFunction implements DataRuleFunction {
 
 		dataRuleFunctionResult.setDataDefinitionField(dataDefinitionField);
 		dataRuleFunctionResult.setErrorCode(
-			DataDefinitionRuleConstants.INVALID_URL_ERROR);
+			DataDefinitionRuleConstants.VALUE_MUST_BE_INTEGER_ERROR);
 		dataRuleFunctionResult.setValid(false);
 
 		if (value == null) {
 			return dataRuleFunctionResult;
 		}
 
-		boolean result = Validator.isUrl(value.toString());
+		Integer valueInteger = NumberUtils.toInt(
+			value.toString(), Integer.MIN_VALUE);
+
+		boolean result = false;
+
+		if (valueInteger != Integer.MIN_VALUE) {
+			result = true;
+		}
 
 		dataRuleFunctionResult.setValid(result);
